Move Search block to top of left column
