+++ server/src/frontend/components/Components/Search/style.css
