+++ server/src/frontend/components/ComponentsTree/style.css
@@ -17,8 +17,3 @@
     justify-content: center;
     align-items: center;
 }
-
-.search-block {
-    padding-top: 10px;
-    padding-bottom: 10px;
-}
